+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Added
+
+- Git commit parsing for TeraChem as part of version parsing
+
+### Changed
+
+- `qcio` `>=0.1.0` -> `>=0.2.0`
+
 ## [0.3.0]
 
 ### Changed